Make thumbnails hold more details.
# Conflicts:
#	src/Generators/ImageThumbnailGenerator.php
#	src/Generators/PDFThumbnailGenerator.php
#	src/Models/Asset.php
