+++ src/Contracts/ThumbnailGeneratorInterface.php
@@ -10,4 +10,11 @@ interface ThumbnailGeneratorInterface
      * Generates a thumbnail url for an asset.
      */
     public function generateThumbnailUrl(Asset $asset): ?string;
+
+    /**
+     * Gets the type of output this generator produces.
+     *
+     * @return string
+     */
+    public function getOutputType(): string;
 }
