+++ src/Generators/ImageThumbnailGenerator.php
@@ -10,6 +10,13 @@ class ImageThumbnailGenerator implements ThumbnailGeneratorInterface
     /**
      * {@inheritdoc}
      */
+    public function getOutputType(): string {
+        return 'image';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function generateThumbnailUrl(Asset $asset): string
     {
         return $asset->url;
